+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 # 2017-12-28
 
 * Added some lib32gcc compiler support installs to ubuConfig script.  Helps with 64-bit operating systems.
+
+* Added some /etc/skel mods to the script, including a by-default .ssh folder and a copy of the doBackups script for all newly created profiles